+++ src/main/java/com/nc/airport/backend/model/entities/model/flight/Flight.java
@@ -8,6 +8,7 @@ import com.nc.airport.backend.eav.annotations.attribute.value.ValueField;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Getter
@@ -22,62 +23,14 @@ public class Flight {
     private LocalDateTime arrivalDatetime;
 
     @ReferenceField(ID = "8")
-    private int airplaneId;
+    private BigInteger airplaneId;
 
     @ValueField(ID = "9")
     private double baseCost;
 
     @ReferenceField(ID = "10")
-    private int arrivalAirport;
+    private BigInteger arrivalAirport;
 
     @ReferenceField(ID = "11")
-    private int departureAirport;
-
-//    public LocalDateTime getDepartureDatetime() {
-//        return departureDatetime;
-//    }
-//
-//    public void setDepartureDatetime(LocalDateTime departureDatetime) {
-//        this.departureDatetime = departureDatetime;
-//    }
-//
-//    public LocalDateTime getArrivalDatetime() {
-//        return arrivalDatetime;
-//    }
-//
-//    public void setArrivalDatetime(LocalDateTime arrivalDatetime) {
-//        this.arrivalDatetime = arrivalDatetime;
-//    }
-//
-//    public int getAirplaneId() {
-//        return airplaneId;
-//    }
-//
-//    public void setAirplaneId(int airplaneId) {
-//        this.airplaneId = airplaneId;
-//    }
-//
-//    public double getBaseCost() {
-//        return baseCost;
-//    }
-//
-//    public void setBaseCost(double baseCost) {
-//        this.baseCost = baseCost;
-//    }
-//
-//    public int getArrivalAirport() {
-//        return arrivalAirport;
-//    }
-//
-//    public void setArrivalAirport(int arrivalAirport) {
-//        this.arrivalAirport = arrivalAirport;
-//    }
-//
-//    public int getDepartureAirport() {
-//        return departureAirport;
-//    }
-//
-//    public void setDepartureAirport(int departureAirport) {
-//        this.departureAirport = departureAirport;
-//    }
+    private BigInteger departureAirport;
 }
